+++ internal/impl/vmware/package.go
@@ -1,0 +1,3 @@
+// Package vmware implements components that collect data from VMware
+// vCenter and ESXi endpoints.
+package vmware