+++ public/components/all/package.go
@@ -42,6 +42,7 @@ import (
 	_ "github.com/warpstreamlabs/bento/public/components/nats"
 	_ "github.com/warpstreamlabs/bento/public/components/nsq"
 	_ "github.com/warpstreamlabs/bento/public/components/opensearch"
+	_ "github.com/warpstreamlabs/bento/public/components/openshift"
 	_ "github.com/warpstreamlabs/bento/public/components/opensnowcat"
 	_ "github.com/warpstreamlabs/bento/public/components/otlp"
 	_ "github.com/warpstreamlabs/bento/public/components/prometheus"
@@ -61,6 +62,7 @@ import (
 	_ "github.com/warpstreamlabs/bento/public/components/sql"
 	_ "github.com/warpstreamlabs/bento/public/components/statsd"
 	_ "github.com/warpstreamlabs/bento/public/components/twitter"
+	_ "github.com/warpstreamlabs/bento/public/components/vmware"
 	_ "github.com/warpstreamlabs/bento/public/components/wasm"
 	_ "github.com/warpstreamlabs/bento/public/components/zeromq"
 )