+++ public/components/openshift/package.go
@@ -1,0 +1,8 @@
+// Package openshift imports all component implementations related to
+// OpenShift.
+package openshift
+
+import (
+	// Bring in the internal plugin definitions.
+	_ "github.com/warpstreamlabs/bento/internal/impl/openshift"
+)